Move axes code to common sub-module
