+++ src/ome_zarr_models/_v05/axes.py
@@ -1,3 +1,3 @@
-from ome_zarr_models.v04.axes import Axes, Axis, AxisType
+from ome_zarr_models.common.axes import Axes, Axis, AxisType
 
 __all__ = ["Axes", "Axis", "AxisType"]
